Bug fix to correctly link pg_hba.conf and postgresql.conf files

+++ postgres/pgs_init.sh
@@ -1,11 +1,12 @@
 #!/bin/bash
 
-if [ ! -L ${PGDATA}/postgresql.conf ]; then
+if [[ -f ${PGDATA}/postgresql.conf ]] && [[ ! -L ${PGDATA}/postgresql.conf ]]; then
     # Force postgresql.conf in datadir to be a link to the default conf file
     # Default conf file is copied into the container at build time
+    # This will update newly-created datadirs.
+    # A similar check in check_datadir.sh will ensure the link for already created datadirs
     ln -sf /etc/postgresql/postgresql.conf ${PGDATA}/postgresql.conf
 fi
-
 ## http://www.postgresql.org/docs/9.1/static/auth-pg-hba-conf.html
 net="0.0.0.0/0"
